[expo-manifests][ios] Rename classes to be more swifty
