+++ packages/expo-manifests/CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ### 💡 Others
 
-- Convert iOS implementation to Swift. ([#21298](https://github.com/expo/expo/pull/21298) by [@wschurman](https://github.com/wschurman))
+- Convert iOS implementation to Swift. ([#21298](https://github.com/expo/expo/pull/21298), [#21648](https://github.com/expo/expo/pull/21648) by [@wschurman](https://github.com/wschurman))
 
 ## 0.5.1 — 2023-02-09
