+++ packages/expo-manifests/ios/Tests/LegacyManifestSpec.swift
@@ -4,15 +4,11 @@ import ExpoModulesTestCore
 
 @testable import EXManifests
 
-enum ManifestTestError: Error {
-    case testError
-}
-
-class EXManifestsBareManifestSpec : ExpoSpec {
+class LegacyManifestSpec : ExpoSpec {
   override func spec() {
     describe("instantiation") {
       it("instantiates and reads properties") {
-        let manifestJson = "{\"id\":\"0eef8214-4833-4089-9dff-b4138a14f196\",\"commitTime\":1609975977832}"
+        let manifestJson = "{\"sdkVersion\":\"39.0.0\",\"id\":\"@esamelson/native-component-list\",\"releaseId\":\"0eef8214-4833-4089-9dff-b4138a14f196\",\"commitTime\":\"2020-11-11T00:17:54.797Z\",\"bundleUrl\":\"https://classic-assets.eascdn.net/%40esamelson%2Fnative-component-list%2F39.0.0%2F01c86fd863cfee878068eebd40f165df-39.0.0-ios.js\"}"
         let manifestData = manifestJson.data(using: .utf8)
         guard let manifestData = manifestData else {
           throw ManifestTestError.testError
@@ -22,20 +18,24 @@ class EXManifestsBareManifestSpec : ExpoSpec {
           throw ManifestTestError.testError
         }
 
-        let manifest = EXManifestsBareManifest(rawManifestJSON: manifestJsonObject)
+        let manifest = LegacyManifest(rawManifestJSON: manifestJsonObject)
 
-        expect(manifest.rawId()) == "0eef8214-4833-4089-9dff-b4138a14f196"
-        expect(manifest.commitTimeNumber()) == 1609975977832
-        expect(manifest.metadata()).to(beNil())
+        expect(manifest.releaseID()) == "0eef8214-4833-4089-9dff-b4138a14f196"
+        expect(manifest.commitTime()) == "2020-11-11T00:17:54.797Z"
+        expect(manifest.bundledAssets()).to(beNil())
+        expect(manifest.runtimeVersion()).to(beNil())
+        expect(manifest.bundleKey()).to(beNil())
+        expect(manifest.assetUrlOverride()).to(beNil())
 
         // from base class
-        expect(manifest.stableLegacyId()) == "0eef8214-4833-4089-9dff-b4138a14f196"
-        expect(manifest.scopeKey()) == "0eef8214-4833-4089-9dff-b4138a14f196"
+        expect(manifest.stableLegacyId()) == "@esamelson/native-component-list"
+        expect(manifest.scopeKey()) == "@esamelson/native-component-list"
         expect(manifest.easProjectId()).to(beNil())
-        expect(manifest.sdkVersion()).to(beNil())
+        expect(manifest.bundleUrl()) == "https://classic-assets.eascdn.net/%40esamelson%2Fnative-component-list%2F39.0.0%2F01c86fd863cfee878068eebd40f165df-39.0.0-ios.js"
+        expect(manifest.sdkVersion()) == "39.0.0"
 
         // from base base class
-        expect(manifest.legacyId()) == "0eef8214-4833-4089-9dff-b4138a14f196"
+        expect(manifest.legacyId()) == "@esamelson/native-component-list"
         expect(manifest.revisionId()).to(beNil())
         expect(manifest.slug()).to(beNil())
         expect(manifest.appKey()).to(beNil())
@@ -62,7 +62,7 @@ class EXManifestsBareManifestSpec : ExpoSpec {
         expect(manifest.iosSplashImageResizeMode()).to(beNil())
         expect(manifest.iosGoogleServicesFile()).to(beNil())
         expect(manifest.supportsRTL()) == false
-        expect(manifest.jsEngine()) == "hermes"
+        expect(manifest.jsEngine()) == "jsc"
       }
     }
   }
