+++ packages/expo-updates/ios/Tests/LegacyUpdateSpec.swift
@@ -19,7 +19,7 @@ class LegacyUpdateSpec : ExpoSpec {
   override func spec() {
     describe("bundledAssetBaseUrl") {
       it("expo domain") {
-        let manifest = EXManifestsLegacyManifest(rawManifestJSON: [:])
+        let manifest = LegacyManifest(rawManifestJSON: [:])
         let expected = URL(string: "https://classic-assets.eascdn.net/~assets/")
         expect(LegacyUpdate.bundledAssetBaseUrl(withManifest: manifest, config: UpdatesConfig.config(fromDictionary: [UpdatesConfig.EXUpdatesConfigUpdateUrlKey: "https://exp.host/@test/test"]))) == expected
         expect(LegacyUpdate.bundledAssetBaseUrl(withManifest: manifest, config: UpdatesConfig.config(fromDictionary: [UpdatesConfig.EXUpdatesConfigUpdateUrlKey: "https://expo.io/@test/test"]))) == expected
@@ -27,7 +27,7 @@ class LegacyUpdateSpec : ExpoSpec {
       }
 
       it("expo subdomain") {
-        let manifest = EXManifestsLegacyManifest(rawManifestJSON: [:])
+        let manifest = LegacyManifest(rawManifestJSON: [:])
         let expected = URL(string: "https://classic-assets.eascdn.net/~assets/")
         expect(LegacyUpdate.bundledAssetBaseUrl(withManifest: manifest, config: UpdatesConfig.config(fromDictionary: [UpdatesConfig.EXUpdatesConfigUpdateUrlKey: "https://staging.exp.host/@test/test"]))) == expected
         expect(LegacyUpdate.bundledAssetBaseUrl(withManifest: manifest, config: UpdatesConfig.config(fromDictionary: [UpdatesConfig.EXUpdatesConfigUpdateUrlKey: "https://staging.expo.io/@test/test"]))) == expected
@@ -36,56 +36,56 @@ class LegacyUpdateSpec : ExpoSpec {
 
       it("AssetUrlOverride AbsoluteUrl") {
         let absoluteUrlString = "https://xxx.dev/~assets"
-        let manifest = EXManifestsLegacyManifest(rawManifestJSON: ["assetUrlOverride": absoluteUrlString])
+        let manifest = LegacyManifest(rawManifestJSON: ["assetUrlOverride": absoluteUrlString])
         let absoluteExpected = URL(string: absoluteUrlString)
         let absoluteActual = LegacyUpdate.bundledAssetBaseUrl(withManifest: manifest, config: self.selfHostedConfig)
         expect(absoluteActual) == absoluteExpected
       }
 
       it("AssetUrlOverride RelativeUrl") {
-        let manifest = EXManifestsLegacyManifest(rawManifestJSON: ["assetUrlOverride": "my_assets"])
+        let manifest = LegacyManifest(rawManifestJSON: ["assetUrlOverride": "my_assets"])
         let relativeExpected = URL(string: "https://esamelson.github.io/self-hosting-test/my_assets")
         let relativeActual = LegacyUpdate.bundledAssetBaseUrl(withManifest: manifest, config: self.selfHostedConfig)
         expect(relativeActual) == relativeExpected
       }
 
       it("AssetUrlOverride OriginRelativeUrl") {
-        let manifest = EXManifestsLegacyManifest(rawManifestJSON: ["assetUrlOverride": "/my_assets"])
+        let manifest = LegacyManifest(rawManifestJSON: ["assetUrlOverride": "/my_assets"])
         let originRelativeExpected = URL(string: "https://esamelson.github.io/my_assets")
         let originRelativeActual = LegacyUpdate.bundledAssetBaseUrl(withManifest: manifest, config: self.selfHostedConfig)
         expect(originRelativeActual) == originRelativeExpected
       }
 
       it("AssetUrlOverride RelativeUrlDotSlash") {
-        let manifest = EXManifestsLegacyManifest(rawManifestJSON: ["assetUrlOverride": "./my_assets"])
+        let manifest = LegacyManifest(rawManifestJSON: ["assetUrlOverride": "./my_assets"])
         let relativeDotSlashExpected = URL(string: "https://esamelson.github.io/self-hosting-test/my_assets")
         let relativeDotSlashActual = LegacyUpdate.bundledAssetBaseUrl(withManifest: manifest, config: self.selfHostedConfig)
         expect(relativeDotSlashActual) == relativeDotSlashExpected
       }
 
       it("AssetUrlOverride Normalize") {
-        let manifest = EXManifestsLegacyManifest(rawManifestJSON: ["assetUrlOverride": "./a/../b"])
+        let manifest = LegacyManifest(rawManifestJSON: ["assetUrlOverride": "./a/../b"])
         let expected = URL(string: "https://esamelson.github.io/self-hosting-test/b")
         let actual = LegacyUpdate.bundledAssetBaseUrl(withManifest: manifest, config: self.selfHostedConfig)
         expect(actual) == expected
       }
 
       it("AssetUrlOverride NormalizeToHostname") {
-        let manifest = EXManifestsLegacyManifest(rawManifestJSON: ["assetUrlOverride": "../b"])
+        let manifest = LegacyManifest(rawManifestJSON: ["assetUrlOverride": "../b"])
         let expected = URL(string: "https://esamelson.github.io/b")
         let actual = LegacyUpdate.bundledAssetBaseUrl(withManifest: manifest, config:self.selfHostedConfig)
         expect(actual) == expected
       }
 
       it("AssetUrlOverride NormalizePastHostname") {
-        let manifest = EXManifestsLegacyManifest(rawManifestJSON: ["assetUrlOverride": "../../b"])
+        let manifest = LegacyManifest(rawManifestJSON: ["assetUrlOverride": "../../b"])
         let expected = URL(string: "https://esamelson.github.io/b")
         let actual = LegacyUpdate.bundledAssetBaseUrl(withManifest: manifest, config:self.selfHostedConfig)
         expect(actual) == expected
       }
 
       it("AssetUrlOverride Default") {
-        let manifest = EXManifestsLegacyManifest(rawManifestJSON: [:])
+        let manifest = LegacyManifest(rawManifestJSON: [:])
         let defaultExpected = URL(string: "https://esamelson.github.io/self-hosting-test/assets")
         let defaultActual = LegacyUpdate.bundledAssetBaseUrl(withManifest: manifest, config:self.selfHostedConfig)
         expect(defaultActual) == defaultExpected
@@ -95,7 +95,7 @@ class LegacyUpdateSpec : ExpoSpec {
     describe("instantiation") {
       it("development") {
         // manifests served from a developer tool should not need the releaseId and commitTime fields
-        let manifest = EXManifestsLegacyManifest(
+        let manifest = LegacyManifest(
           rawManifestJSON: [
             "sdkVersion": "39.0.0",
             "bundleUrl": "https://url.to/bundle.js",
@@ -109,7 +109,7 @@ class LegacyUpdateSpec : ExpoSpec {
       
       it("production all fields") {
         // production manifests should require the releaseId, commitTime, sdkVersion, and bundleUrl fields
-        let manifest = EXManifestsLegacyManifest(
+        let manifest = LegacyManifest(
           rawManifestJSON: [
             "sdkVersion": "39.0.0",
             "releaseId": "0eef8214-4833-4089-9dff-b4138a14f196",
@@ -121,7 +121,7 @@ class LegacyUpdateSpec : ExpoSpec {
       }
       
       it("production no sdk version") {
-        let manifest = EXManifestsLegacyManifest(
+        let manifest = LegacyManifest(
           rawManifestJSON: [
             "releaseId": "0eef8214-4833-4089-9dff-b4138a14f196",
             "commitTime": "2020-11-11T00:17:54.797Z",
@@ -132,7 +132,7 @@ class LegacyUpdateSpec : ExpoSpec {
       }
       
       it("production no releaseId") {
-        let manifest = EXManifestsLegacyManifest(
+        let manifest = LegacyManifest(
           rawManifestJSON: [
             "sdkVersion": "39.0.0",
             "commitTime": "2020-11-11T00:17:54.797Z",
@@ -143,7 +143,7 @@ class LegacyUpdateSpec : ExpoSpec {
       }
       
       it("production no commitTime") {
-        let manifest = EXManifestsLegacyManifest(
+        let manifest = LegacyManifest(
           rawManifestJSON: [
             "sdkVersion": "39.0.0",
             "releaseId": "0eef8214-4833-4089-9dff-b4138a14f196",
@@ -154,7 +154,7 @@ class LegacyUpdateSpec : ExpoSpec {
       }
       
       it("production no bundleUrl") {
-        let manifest = EXManifestsLegacyManifest(
+        let manifest = LegacyManifest(
           rawManifestJSON: [
             "sdkVersion": "39.0.0",
             "releaseId": "0eef8214-4833-4089-9dff-b4138a14f196",
@@ -166,7 +166,7 @@ class LegacyUpdateSpec : ExpoSpec {
       
       it("setsUpdateRuntimeAsSdkIfNoManifestRuntime") {
         let sdkVersion = "39.0.0"
-        let manifest = EXManifestsLegacyManifest(
+        let manifest = LegacyManifest(
           rawManifestJSON: [
             "sdkVersion": sdkVersion,
             "releaseId": "0eef8214-4833-4089-9dff-b4138a14f196",
@@ -181,7 +181,7 @@ class LegacyUpdateSpec : ExpoSpec {
       it("setsUpdateRuntimeAsRuntimeIfBothManifestRuntime") {
         let sdkVersion = "39.0.0"
         let runtimeVersion = "hello"
-        let manifest = EXManifestsLegacyManifest(
+        let manifest = LegacyManifest(
           rawManifestJSON: [
             "runtimeVersion": runtimeVersion,
             "sdkVersion": sdkVersion,
